requirements.txt
pylint==2.17.4
pytest
voluptuous==0.13.1
pytest-voluptuous== 1.2.0
python-dotenv==0.21.1
requests
mimesis
black
pydantic